# Optimize resizing process via caching: 204ms to 2ms response time.