@amyboyd 2015-11-26
Here is the result of hitting the API with 100000 requests, 500 requests at a time, over 3.5 seconds. The average response time after this PR is still only 17ms.

$ ab -c 500 -n 100000 'http://placingbad.dev/123/123/walter'
This is ApacheBench, Version 2.3 <$Revision: 1528965 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking placingbad.dev (be patient)
Completed 10000 requests
Completed 20000 requests
Completed 30000 requests
Completed 40000 requests
Completed 50000 requests
Completed 60000 requests
Completed 70000 requests
Completed 80000 requests
Completed 90000 requests
Completed 100000 requests
Finished 100000 requests


Server Software:        nginx
Server Hostname:        placingbad.dev
Server Port:            80

Document Path:          /123/123/walter
Document Length:        166 bytes

Concurrency Level:      500
Time taken for tests:   3.574 seconds
Complete requests:      100000
Failed requests:        4253
   (Connect: 0, Receive: 0, Length: 4253, Exceptions: 0)
Non-2xx responses:      95747
Total transferred:      50697824 bytes
HTML transferred:       35785283 bytes
Requests per second:    27978.02 [#/sec] (mean)
Time per request:       17.871 [ms] (mean)
Time per request:       0.036 [ms] (mean, across all concurrent requests)
Transfer rate:          13851.81 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    7  69.1      2    1005
Processing:     0    9  22.3      3     204
Waiting:        0    8  22.3      3     203
Total:          1   15  72.8      5    1133

Percentage of the requests served within a certain time (ms)
  50%      5
  66%      5
  75%      7
  80%      8
  90%     13
  95%     18
  98%    120
  99%    128
 100%   1133 (longest request)